requirements.txt
aiohttp==3.6.0
async-timeout==3.0.1
certifi==2018.11.29
chardet==3.0.4
discord.py==1.3
idna==2.8
multidict==4.5.2
nba-api==1.1.8
objectpath==0.6.1
python-dotenv==0.10.1
requests==2.21.0
urllib3==1.24.2
websockets==6.0
